--- src/Textfeild/index.js
@@ -1,2 +1,0 @@
-import { PSATextfeild } from "./textfeild"; 
-export default PSATextfeild;
