+++ src/Textfield/index.js
@@ -1,0 +1,2 @@
+import { PSATextfield } from "./textfield";
+export default PSATextfield;
