+++ src/index.js
@@ -3,3 +3,4 @@ export { default as Spinner } from "./spinner/index";
 export { default as Map } from "./map/index";
 export { default as FadeAlert } from "./fade-alert/index";
 export { default as NcalcMap } from "./ncalc-map/index";
+export { default as TextField } from "./Textfield/index";
